Update to be SwiftUI-first
